+++ studio/treeComponents/ConsoleTreeComponent.js
@@ -13,6 +13,10 @@ class ConsoleTreeComponent extends AbstractTreeComponent {
     return Math.floor((this.getRootNode().getBodyShadowDimensions().height - 60) * 0.3)
   }
 
+  setFile(fileName) {
+    this.setWord(1, fileName)
+  }
+
   toHakonCode() {
     return `.consoleOutput
  height ${this._getHeight()}px
@@ -24,12 +28,14 @@ class ConsoleTreeComponent extends AbstractTreeComponent {
   }
 
   _getMessageBuffer() {
+    // todo: cleanup
     const app = this.getRootNode()
     const tab = app.getMountedTab()
     return tab ? tab.getMessageBuffer() : app.getMessageBuffer()
   }
 
   getDependencies() {
+    // todo: cleanup
     // 2 dependencies. the program and the programs message buffer.
     // let's call the latter the panel buffer for now.
     const deps = this.getParent().getDependencies()