+++ studio/treeComponents/HelpModal.js
@@ -95,14 +95,11 @@ p
 p Current working folder: ${app.getDefaultDisk().getPathBase()}
 p Version ${app.getVersion()} ${app.constructor.name}
 p
- a Welcome Page
+ a Open Demo Page
   id welcomePageButton
   clickCommand openOhayoProgramCommand
   value ohayo.ohayo
-a Keyboard Shortcuts
- class helpToggle
- clickCommand toggleShadowByIdCommand
- value shortcutsHelp
+div Keyboard Shortcuts:
 ${this._getShortcutsHelpStumpCode()}`
   }
 }