+++ studio/treeComponents/MenuTreeComponent.js
@@ -1,27 +1,53 @@
 const { jtree } = require("jtree")
 const { AbstractTreeComponent } = require("jtree/products/TreeComponentFramework.node.js")
 
-const NewDropDownMenuTreeComponent = require("./NewDropDownMenuTreeComponent.js")
 const StudioConstants = require("./StudioConstants.js")
+const { TabsTreeComponent } = require("./TabsTreeComponent.js")
+
+class LogoTreeComponent extends AbstractTreeComponent {
+  toStumpCode() {
+    return `a help
+ clickCommand toggleHelpCommand
+ class LogoTreeComponent`
+  }
+}
+
+class NewButtonTreeComponent extends AbstractTreeComponent {
+  toStumpCode() {
+    return `a &nbsp;+
+ id newButton
+ class NewButtonTreeComponent
+ clickCommand createNewBlankProgramCommand
+ value untitled.ohayo`
+  }
+}
 
 class MenuTreeComponent extends AbstractTreeComponent {
   createParser() {
     return new jtree.TreeNode.Parser(undefined, {
-      newDropDownMenu: NewDropDownMenuTreeComponent
+      logo: LogoTreeComponent,
+      tabs: TabsTreeComponent,
+      newButton: NewButtonTreeComponent
     })
   }
 
-  getDependencies() {
-    return [{ getLineModifiedTime: () => this.getParent().getWindowSizeMTime() }]
+  toggleVisibility() {
+    this.setWord(1, this.isVisible() ? "hidden" : "visible")
+  }
+
+  isVisible() {
+    return this.getWord(1) === "visible"
   }
 
   toHakonCode() {
     const theme = this.getTheme()
+    const display = this.isVisible() ? "flex" : "none"
     return `.MenuTreeComponent
  ${theme.disableTextSelect(1)}
  font-size 14px
  padding-left 5px
  box-sizing border-box
+ overflow-x scroll
  right 0
  left 0
  position relative
@@ -29,23 +55,12 @@ class MenuTreeComponent extends AbstractTreeComponent {
  z-index 92
  white-space nowrap
  background ${theme.menuBackground}
- color ${theme.darkBlack}
- display flex
- a
+ display ${display}
+ .LogoTreeComponent,.NewButtonTreeComponent
   padding-right 5px
   line-height 30px
   display inline-block
   color ${theme.menuTreeComponentColor}`
-  }
-
-  toStumpCode() {
-    return `div
- class MenuTreeComponent ${this.constructor.name}
- a ${StudioConstants.productName}
-  clickCommand toggleHelpCommand
- a New ▾
-  id newToggle
-  clickCommand toggleAndRenderNewDropDownCommand`
   }
 }